Macro solana_sdk_macro::respan

respan!() { /* proc-macro */ }

A proc-macro which respans the tokens in its first argument (a Path) to be resolved at the tokens of its second argument. For internal use only.

There must be exactly one comma in the input, which is used to separate the two arguments. The second argument should be exactly one token.

For example, respan!($crate::foo, with_span) produces the tokens $crate::foo, but resolved at the span of with_span.

The input to this function should be very short - its only purpose is to override the span of a token sequence containing $crate. For all other purposes, a more general proc-macro should be used.